Use standard float for calculations, even if a float16 is passed

+++ stingray/simulator/base.py
@@ -125,7 +125,7 @@ def simulate_times_from_count_array(time, counts, gti, dt, use_spline=False):
     ValueError: simulate_times can only work with...
     """
     time = np.asarray(time)
-    counts = np.asarray(counts)
+    counts = np.asarray(counts).astype(float)
     gti = np.asarray(gti)
     kind = "linear"
     if use_spline and time.size > 2:
